feature(cleaning): one function has been moved to avoid duplication
