+++ mlexplainer/utils/data_processing.py
@@ -1,6 +1,21 @@
 from pandas import concat, DataFrame, Series, merge
 
 
+def calculate_min_max_value(dataframe: DataFrame, feature: str):
+    """
+    Calculate the minimum and maximum values of a feature in a DataFrame.
+    Args:
+        dataframe (DataFrame): The DataFrame containing the feature.
+        feature (str): The name of the feature to calculate min and max values.
+    Returns:
+        tuple: A tuple containing the minimum and maximum values of the feature.
+    """
+    if dataframe[feature].dtype == "category":
+        return 0, dataframe[feature].value_counts().shape[0] - 1
+
+    return dataframe[feature].min(), dataframe[feature].max()
+
+
 def get_index(column_name: str, dataframe: DataFrame) -> int:
     """Extract the index of a column in a DataFrame.
